requirements.txt
dspy
openai
python-dotenv
pandas
datasets
scikit-learn
sentence-transformers